fix(plan): retire the [verbatim] AC tag contract
nax held three incompatible positions on the same tag:

  verbatim-fidelity.ts  spec ACs tagged [verbatim] MUST survive
                        character-for-character; on a miss, fire a repair
                        turn that restores them exactly
  plan-builder.ts:263   instruct the planner to REMOVE [verbatim] tags and
                        rewrite the AC as a behavioural assertion
  spec-drift.ts         treat a leading [verbatim] tag as drift to flag

So the planner was told to strip the tag, the fidelity check then reported
those ACs as dropped, the repair prompt restored them, and spec-drift
flagged the result. The machinery actively fought the rewrite it asked for.

What [verbatim] protects is the reason to drop it, not keep it. Its own
header says it preserves "executable greps, file-existence checks,
regex/count assertions" -- and every real usage in the corpus is a
file-content assertion ("pyproject.toml contains the substring litellm").
Those pass when the string appears anywhere, including a comment, prove
nothing about importability or wiring, and cannot be written as a
fail-first runtime test. An AC is a behavioural test case; the "drop" the
check detected was the planner correctly converting one.

Already dormant in practice: no spec has used the tag since 2026-06-05,
and 48/12585 PRD ACs (0.4%) still carry it.

Removed: verbatim-fidelity.ts and its barrel export, buildVerbatimRepair,
the plan-refine self-heal step and its disk-read detector,
warnOnDroppedVerbatimAcs and its three call sites (plan / plan-refine /
debate), plus the tag-preservation clauses in SPEC_ANCHOR_RULES rule 1 and
the plan-builder audit step -- both now say to rewrite a deprecated-tag AC
into the behaviour it was meant to prove.

Kept: spec-drift's deprecated-tag flag, so a legacy spec still gets
reported, and plan-builder's strip-and-convert instruction.

verbatim-warn.ts held three functions but only one was verbatim-related;
renamed to plan-fidelity.ts for out-of-scope backfill + spec-drift warning.
Test helper verbatim-warn-spy.ts likewise generalised to warn-spy.ts --
withWarnSpy is still used by the out-of-scope tests; only the verbatimWarn
selector died.

Two integration tests were asserting on removed config keys as incidental
fixtures and are retargeted to live ones (quality.autofix.enabled for the
boolean-diff case, precheck.storySizeGate.maxAcCount for the nested
description case).

typecheck, lint, 10092 unit and 1090 integration tests all green.
