+++ src/operations/plan.ts
@@ -6,8 +6,8 @@ import { validatePlanOutput } from "../prd/schema";
 import type { PRD } from "../prd/types";
 import { PlanPromptBuilder } from "../prompts";
 import type { PackageSummary } from "../prompts";
+import { backfillOutOfScope } from "./plan-fidelity";
 import type { RunOperation } from "./types";
-import { backfillOutOfScope, warnOnDroppedVerbatimAcs } from "./verbatim-warn";
 
 export interface PlanInteractiveInput {
   specContent: string;
@@ -74,9 +74,6 @@ export const planInteractiveOp: RunOperation<PlanInteractiveInput, PRD, PlanConf
   },
   verify: async (parsed, input, _ctx) => {
     if (!parsed.userStories || parsed.userStories.length === 0) return null;
-    // Single mode is one-shot (no self-heal turn) — warn on residual [verbatim]
-    // drift and continue. Shared with refine via warnOnDroppedVerbatimAcs.
-    warnOnDroppedVerbatimAcs(parsed, input.specContent, input.featureName);
     // Feature-level exclusions have one home, so a drop is repairable rather
     // than merely reportable — backfill instead of warning-and-continuing.
     return backfillOutOfScope(parsed, input.specContent, input.featureName);
