--- src/operations/verbatim-warn.ts
@@ -1,69 +1,0 @@
-import { getSafeLogger } from "../logger";
-import {
-  applyOutOfScopeFallback,
-  findMissingOutOfScope,
-  findMissingVerbatimAcs,
-  findSpecDriftViolations,
-} from "../prd";
-import type { PRD } from "../prd/types";
-
-/**
- * Soft gate shared by the plan ops (single + refine): warn — loudly, structured
- * — when a `[verbatim]` spec AC did not survive into the PRD. Paraphrasing a
- * verbatim grep / file-check / invariant destroys its verification mechanism
- * (docs/findings/nax-plan-prd-fidelity.md), but `nax plan` is intentionally
- * recovery-tolerant: it always produces a usable PRD rather than failing. The
- * warning is the residual-drift signal, and spec-review remains the explicit
- * gate before any story executes. Refine additionally attempts a same-session
- * self-heal turn before this check; single is one-shot, so this warning is its
- * only verbatim signal.
- */
-export function warnOnDroppedVerbatimAcs(prd: PRD, specContent: string, featureName: string): void {
-  const missing = findMissingVerbatimAcs(specContent, prd);
-  if (missing.length > 0) {
-    getSafeLogger()?.warn(
-      "plan",
-      "[verbatim] spec acceptance criteria dropped from PRD — run spec-review --prd before executing",
-      { featureName, missingCount: missing.length, missing },
-    );
-  }
-}
-
-/**
- * Scope-fidelity backfill shared by the plan ops (single + refine).
- *
- * Unlike the `[verbatim]` gate — which can only warn, because restoring an AC
- * requires knowing which story owns it — feature-level exclusions have exactly
- * one home (`prd.outOfScope`), so a dropped item can be repaired
- * deterministically. The prompt asks the planner for its own wording; whatever
- * it drops is appended verbatim from the spec here, and the warning records
- * that the planner needed the safety net.
- *
- * Returns the backfilled PRD (the input reference when nothing was missing).
- */
-export function backfillOutOfScope(prd: PRD, specContent: string, featureName: string): PRD {
-  const missing = findMissingOutOfScope(specContent, prd);
-  if (missing.length === 0) return prd;
-  getSafeLogger()?.warn("plan", "Spec out-of-scope statements dropped from PRD — backfilled verbatim from the spec", {
-    featureName,
-    missingCount: missing.length,
-    missing,
-  });
-  return applyOutOfScopeFallback(prd, specContent);
-}
-
-/**
- * Residual-drift warning for spec-guard: fires when the specGuard repair turn
- * did not eliminate all behavioral-fidelity violations. Non-fatal — the plan
- * continues with a warning so the user can manually correct or rerun.
- */
-export function warnOnSpecDrift(prd: PRD, featureName: string): void {
-  const violations = findSpecDriftViolations(prd);
-  if (violations.length > 0) {
-    getSafeLogger()?.warn("plan", "spec-drift violations remain after specGuard repair — review PRD before executing", {
-      featureName,
-      violationCount: violations.length,
-      violations,
-    });
-  }
-}
