+++ test/integration/cli/cli-config-command.test.ts
@@ -190,7 +190,7 @@ describe("Config Command", () => {
 
       expect(output).toContain("Enable automatic agent selection and escalation");
       expect(output).toContain("# Max iterations per feature run");
-      expect(output).toContain("# Require typecheck to pass");
+      expect(output).toContain("# Max acceptance criteria count before flagging");
       expect(output).toContain("# TDD strategy: auto | strict | lite | off");
     });
