+++ test/unit/operations/plan-refine.test.ts
@@ -6,7 +6,7 @@ import { PlanPromptBuilder } from "@/prompts";
 import { planInteractiveOp } from "@/operations";
 import type { VerifyContext } from "@/operations";
 import type { NaxRuntime } from "@/runtime";
-import { makeMockAgentManager, makeSessionManager, makeTestRuntime, verbatimWarn, withTempDir, withWarnSpy } from "@test/helpers";
+import { makeMockAgentManager, makeSessionManager, makeTestRuntime, withTempDir, withWarnSpy } from "@test/helpers";
 import type { HopBodyContext } from "@/operations/types";
 
 const createdRuntimes: NaxRuntime[] = [];
@@ -290,184 +290,8 @@ describe("planRefineOp.hopBody()", () => {
 });
 
 
-describe("planRefineOp — warn-and-continue when self-heal still drops a verbatim AC", () => {
-  test("callOp returns the PRD and warns (does not fail) after the repair turn still misses", async () => {
-    await withWarnSpy(async (warnSpy) => {
-      await withTempDir(async (tempDir) => {
-        const outputPath = join(tempDir, "prd.json");
-        // Every turn (draft, refine, repair) writes a structurally valid PRD that
-        // never contains the [verbatim] grep — the self-heal cannot recover it, so
-        // the run continues but warns.
-        const lackingPrd = makeValidPrd("f", "feat/f");
 
-        const agentManager = makeMockAgentManager({
-          runWithFallbackFn: async (req: AgentRunRequest) => {
-            const hopResult = await req.executeHop!("claude", undefined, { kind: "primary" }, req.runOptions);
-            return { result: { ...hopResult.result, agentFallbacks: [] }, fallbacks: [] };
-          },
-          runAsSessionFn: async () => {
-            await Bun.write(outputPath, JSON.stringify(lackingPrd));
-            return {
-              output: "written",
-              estimatedCostUsd: 1,
-              internalRoundTrips: 1,
-              tokenUsage: { inputTokens: 1, outputTokens: 1 },
-            };
-          },
-        });
 
-        const runtime = makeTestRuntime({ agentManager, sessionManager: makeSessionManager() });
-        createdRuntimes.push(runtime);
-
-        const result = await callOp(
-          {
-            runtime,
-            packageView: runtime.packages.repo(),
-            packageDir: tempDir,
-            agentName: runtime.agentManager.getDefault(),
-            storyId: "f",
-            featureName: "f",
-          },
-          planRefineOp,
-          {
-            specContent: '## ACs\n- [verbatim] `grep -rn "X" src/` returns zero matches',
-            codebaseContext: "",
-            featureName: "f",
-            branchName: "feat/f",
-            outputPath,
-          },
-        );
-
-        expect(result.userStories.length).toBeGreaterThan(0); // plan continues
-        const warn = verbatimWarn(warnSpy);
-        expect(warn).toBeDefined();
-        expect((warn?.[2] as Record<string, unknown>).missingCount).toBe(1);
-      });
-    });
-  });
-});
-
-describe("planRefineOp.verify — [verbatim] residual-drift warning", () => {
-  const SPEC_WITH_VERBATIM = '## Acceptance Criteria\n- [verbatim] `grep -rn "oldSym" src/` returns zero matches';
-
-  function makeVerifyCtx(): VerifyContext<unknown> {
-    const runtime = makeTestRuntime();
-    createdRuntimes.push(runtime);
-    const view = runtime.packages.repo();
-    return {
-      packageView: view,
-      config: view.select(planInteractiveOp.config),
-      readFile: async () => null,
-      fileExists: async () => false,
-    };
-  }
-
-  const input = {
-    specContent: SPEC_WITH_VERBATIM,
-    codebaseContext: "",
-    featureName: "f",
-    branchName: "feat/f",
-    outputPath: "/tmp/x.json",
-  };
-
-  test("warns and still returns the PRD when a [verbatim] spec AC is dropped", async () => {
-    await withWarnSpy(async (warnSpy) => {
-      const prd = makeValidPrd("f", "feat/f"); // default AC does not contain the grep command
-      const result = await planRefineOp.verify?.(prd as never, input as never, makeVerifyCtx());
-      expect(result).toBeTruthy(); // continues
-      const warn = verbatimWarn(warnSpy);
-      expect(warn).toBeDefined();
-      expect((warn?.[2] as Record<string, unknown>).missingCount).toBe(1);
-    });
-  });
-
-  test("does not warn when the [verbatim] command survives in some PRD AC", async () => {
-    await withWarnSpy(async (warnSpy) => {
-      const base = makeValidPrd("f", "feat/f");
-      const prd = {
-        ...base,
-        userStories: [
-          {
-            ...base.userStories[0],
-            acceptanceCriteria: [
-              'When cleanup completes, grep -rn "oldSym" src/ returns zero matches.',
-              "handler rejects invalid input", // satisfies the negative-path structural check
-            ],
-          },
-        ],
-      };
-      const result = await planRefineOp.verify?.(prd as never, input as never, makeVerifyCtx());
-      expect(result).toBeTruthy();
-      expect(verbatimWarn(warnSpy)).toBeUndefined();
-    });
-  });
-});
-
-describe("planRefineOp.hopBody — [verbatim] self-heal turn", () => {
-  const SPEC = '## ACs\n- [verbatim] `grep -rn "X" src/` returns zero matches';
-
-  function turn(output: string, cost: number) {
-    return { output, estimatedCostUsd: cost, internalRoundTrips: 1, tokenUsage: { inputTokens: 0, outputTokens: 0 } };
-  }
-
-  function makeCtx() {
-    const sendWithParseRetry = mock(async () => turn("draft", 1));
-    let sendCount = 0;
-    const send = mock(async (_p: string) => {
-      sendCount += 1;
-      return turn(sendCount === 1 ? "refined" : "repaired", 2);
-    });
-    const ctx = {
-      input: { specContent: SPEC, codebaseContext: "", featureName: "f", branchName: "feat/f", outputPath: "/tmp/p.json" },
-      send,
-      sendWithParseRetry,
-    } as unknown as Parameters<NonNullable<typeof planRefineOp.hopBody>>[1];
-    return { ctx, send, sendWithParseRetry };
-  }
-
-  test("fires exactly one repair turn when the written PRD dropped a verbatim AC", async () => {
-    _planRefineDeps.readFile = async () => JSON.stringify(makeValidPrd("f", "feat/f")); // AC lacks `grep ... "X"`
-    const repairSpy = spyOn(PlanPromptBuilder.prototype, "buildVerbatimRepair").mockReturnValue("REPAIR-PROMPT");
-    const { ctx, send } = makeCtx();
-
-    const result = await planRefineOp.hopBody!("init", ctx);
-
-    expect(send).toHaveBeenCalledTimes(2); // refine + repair
-    expect(repairSpy).toHaveBeenCalledTimes(1);
-    expect(send.mock.calls[1][0]).toBe("REPAIR-PROMPT");
-    expect(result.output).toBe("repaired");
-    expect(result.estimatedCostUsd).toBe(5); // 1 (draft) + 2 (refine) + 2 (repair)
-  });
-
-  test("no repair turn when the written PRD preserved the verbatim AC", async () => {
-    const base = makeValidPrd("f", "feat/f");
-    const preserved = {
-      ...base,
-      userStories: [{ ...base.userStories[0], acceptanceCriteria: ['grep -rn "X" src/ returns zero matches'] }],
-    };
-    _planRefineDeps.readFile = async () => JSON.stringify(preserved);
-    const repairSpy = spyOn(PlanPromptBuilder.prototype, "buildVerbatimRepair");
-    const { ctx, send } = makeCtx();
-
-    const result = await planRefineOp.hopBody!("init", ctx);
-
-    expect(send).toHaveBeenCalledTimes(1); // refine only
-    expect(repairSpy).not.toHaveBeenCalled();
-    expect(result.output).toBe("refined");
-    expect(result.estimatedCostUsd).toBe(3); // 1 (draft) + 2 (refine)
-  });
-
-  test("no repair turn when the PRD file is absent or unparseable", async () => {
-    _planRefineDeps.readFile = async () => null;
-    const repairSpy = spyOn(PlanPromptBuilder.prototype, "buildVerbatimRepair");
-    const { ctx, send } = makeCtx();
-
-    await planRefineOp.hopBody!("init", ctx);
-
-    expect(send).toHaveBeenCalledTimes(1);
-    expect(repairSpy).not.toHaveBeenCalled();
-  });
-});
 
 describe("planRefineOp.hopBody — specGuard spec-drift repair turn", () => {
   const SPEC = "# Spec\n- [unit] does a thing";
